+++ backend/src/ProEventos.Persistence/Contracts/IEventPersist.cs
@@ -1,0 +1,11 @@
+using ProEventos.Domain;
+
+namespace ProEventos.Persistence.Contracts
+{
+    public interface IEventPersist
+    {
+        Task<Event[]> GetAllEventsByThemeAsync(string theme, bool includeSpeakers);
+        Task<Event[]> GetAllEventsAsync(string theme, bool includeSpeakers);
+        Task<Event> GetEventByIdAsync(int eventId, bool includeSpeakers);
+    }
+}
