+++ backend/src/ProEventos.Persistence/Contracts/IGeneralPersist.cs
@@ -1,0 +1,11 @@
+namespace ProEventos.Persistence.Contracts
+{
+    public interface IGeneralPersist
+    {
+        void Add<T>(T entity) where T : class;
+        void Update<T>(T entity) where T : class;
+        void Delete<T>(T entity) where T : class;
+        void DeleteRange<T>(T[] entityArray) where T : class;
+        Task<bool> SaveChangesAsync();
+    }
+}
