+++ backend/src/ProEventos.Persistence/Contracts/ISpeakerPersist.cs
@@ -1,0 +1,11 @@
+using ProEventos.Domain;
+
+namespace ProEventos.Persistence.Contracts
+{
+    public interface ISpeakerPersist
+    {
+        Task<Speaker[]> GetAllSpeakersByNameAsync(string name, bool includeEvents);
+        Task<Speaker[]> GetAllSpeakersAsync(bool includeEvents);
+        Task<Speaker> GetSpeakerByIdAsync(int speakerId, bool includeEvents);
+    }
+}
