+++ backend/src/ProEventos.Persistence/GeralPersist.cs
@@ -1,0 +1,42 @@
+using Microsoft.EntityFrameworkCore;
+using ProEventos.Domain;
+using ProEventos.Persistence.Contracts;
+
+namespace ProEventos.Persistence
+{
+    public class GeralPersist : IGeneralPersist
+    {
+        private readonly ProEventosContext _context;
+
+        public GeralPersist(ProEventosContext context)
+        {
+            _context = context;
+        }
+
+        public void Add<T>(T entity) where T : class
+        {
+            _context.Add(entity);
+        }
+
+        public void Update<T>(T entity) where T : class
+        {
+            _context.Update(entity);
+        }
+
+        public void Delete<T>(T entity) where T : class
+        {
+            _context.Remove(entity);
+        }
+
+        public void DeleteRange<T>(T[] entityArray) where T : class
+        {
+            _context.RemoveRange(entityArray);
+        }
+
+        public async Task<bool> SaveChangesAsync()
+        {
+            return await _context.SaveChangesAsync() > 0;
+        }
+
+    }
+}
